Move Keycloak client
